Update patch set 1
Patch Set 1:

I'd be careful here. According to the release notes (https://pip.pypa.io/en/stable/news) "Default to doing a user install (as if --user was passed) when the main site-packages directory is not writeable and user site-packages are enabled."

This does not necessarily mean --user is always default. It means if the usual system level pip install path is not writeable then fallback to `--user`. Which is very much different than --user being the default IMO as there are specific conditions implied here.

My suggestion here is to be explicit to clarify any potential confusions. If our intention is to use the --user path then explicitly pass --user is the better option rather than relying on a conditional default statement.

Patch-set: 1
CC: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>