Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > sorry should have been more clear:
> > my comment here: https://github.com/pypa/pip/issues/1668#issuecomment-577371128
> > and this is the failure we are seeing in multiple places. https://jira.linuxfoundation.org/projects/IT/queues/custom/74/IT-18779
> 
> Thanks, sounds like a bad bug from upstream.

I don't think it is a bug. I think they are trying to restrict the number of ways for the user to shoot themselves in the foot when using 'user' venvs. I think a 'user' venv can effect a python executable for a given user, even if ~/.venv/bin is not in the PATH.

Patch-set: 1