Remove login shell to fix pip on ubuntu
pip with 20.0.1 works on ubuntu with the login shell removed.

tox-install.sh Failure is:
ERROR: Can not perform a '--user' install. User site-packages are not
visible in this virtualenv.

path with -l
/home/jenkins/.local/bin:/home/jenkins/.local/bin:/usr/local/sbin:\
/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:\
/usr/local/games:/snap/bin:/opt/puppetlabs/bin

path wihout -l
/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:\
/usr/games:/usr/local/games

https://github.com/pypa/pip/issues/5702#issuecomment-524850335
specifically With #7155, pip install --user inside a venv now throws the
expected error "Can not perform a --user install. ..."

but I can't see why we were triggering that error, as we never activate
the venv ...

Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: I44ac3069ecc3b3ff05e572e8cb0e091c6f2607aa