Update patch set 2
Patch Set 2:

I was (foolishly) hoping that the version of PIP would also be shown by pip freeze.  Alas it is not.  Looks like I need to add the following invocation to both python-tools-install.sh and tox-install.sh:  

    python3 -m pip --version

I'll be glad to create the change set, but please advise if you know a better way, if I'm missing something, etc.  Thanks in advance.

Patch-set: 2