Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I was (foolishly) hoping that the version of PIP would also be shown by pip freeze.  Alas it is not.  Looks like I need to add the following invocation to both python-tools-install.sh and tox-install.sh:  
> 
>     python3 -m pip --version
> 
> I'll be glad to create the change set, but please advise if you know a better way, if I'm missing something, etc.  Thanks in advance.

That's how I would do it. Probably worth doing these 3 commands together:

    python --version
    python -m pip --version
    python -m pip freeze

Another useful one that is good to know but I wouldn't put it in a job (mainly because we've been burned by upstream a few times in the past due to incompatability) is `pipdeptree` which is a better version of pip freeze which shows you the chain of package dependencies.

Patch-set: 2