Update patch set 1
Patch Set 1:

Is this meant to make it work on merge commits?
In case of ONAP, in particular, we discourage the creation of merge commits and try to promote cherry-picks instead. Unless we are talking about a new code bring up for other projects. 

When doing a merge commit even, I don't think is a good idea to bring in release files since already merged release files will not make sense to run them through the process.

Thanks!

Patch-set: 1
CC: Gerrit User 65 <65@f2a2391d-0403-4fa5-b183-4056999a8a09>