Update patch set 1
Patch Set 1:

Existing gerrit merge job templates are split on this issue.  For example, the docker and python merge jobs use choosing strategy "gerrit", which is correct in my opinion.  But the well-established maven merge job gerrit-maven-merge uses "default", as does the gerrit-maven-docker-merge job.  Both are wrong in my opinion, but probably don't cause pain very often.

Then there's the maven stage jobs, like gerrit-maven-stage.  This also uses choosing strategy "default" which means a version is built from the tip of the branch and staged, not built from the commit that was merged and referenced in the gerrit review.  Apparently this does not cause pain often either?  I'm just saying, I was really surprised to see that the stage job is like this.

What do you think?

Patch-set: 1