Update patch set 1
Patch Set 1:

In a quiet git repository and reliable CI environment, the choosing strategy 'default' for merge and stage jobs works fine. After merge of a change set, the job runs and succeeds (never needs a re-do). And before any other commits happens, the team stages an image for release, and that job succeeds also.  So in the sunny day scenario, all is well.  

I discovered the need for this change in a busy repository with an unreliable CI environment :/  sound familiar?

Patch-set: 1