Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thank you for the quick reply.  Please tell me more about why merge and stage jobs must always build from tip of the target branch, rather than the commit that was merged.  I need to understand if current Python job templates need to be updated. The tox and pypi merge templates both use choosing-strategy gerrit.  I work with people who rely heavily on those templates in Acumos and ORAN and, to the best of my knowledge, that choice has not (yet) caused problems.

The expectation of our artifact storage is that what's being built is always the latest of a particular tip. This is especially important with maven snapshot artifacts as they get a timestamp encoded into the actual artifact name.

Let's take the follow for instance into account:

Busy maven project has a lot of out of order merges happen and another project that is currently being actively developed in tandem and therefore using snapshot versions instead of release versions (yes this happens, though not nearly as often as it used to thanks to our moving towards self-service releases!)

Downstream project files a bug against the project and it gets merged, but the merge job fails for some reason. In the mean time, a different change, that happens _earlier_ in commit history gets merged, or the merge job is retriggered, in either case, what ends up happening is that the downstream project will end up getting a bad version of the snapshot artifact when they are in fact expecting a fixed version.

As such, our policy has always, and will always be, that a code _merge_ (not the release merge) should only _ever_ build from the tip. By that logic the staging jobs should do the same thing.

Release merge jobs are a special case in that we really do care about point in time and state of the repository for the job, or at least for that one special file we're dealing with.

Patch-set: 1