Python merge use git choosing strategy default
Use git choosing strategy default in tox and pypi merge jobs.
This makes those jobs consistent with maven and other merge
jobs that always build from tip of the target branch to
create artifacts from the latest & greatest code.  Building
from tip (not from Gerrit commit/merge point) avoids confusion
about content when changes are merged out of order.  For
example, a fix is submitted and merged, but the merge job
fails. In the mean time, a different change that happened
earlier in commit history gets merged (or the merge job is
retriggered), causing a new artifact to be pushed.  But that
artifact does not have the expected fix.

Update the github release jobs to match the gerrit jobs.

Document the git commit choosing strategy for the release merge jobs.

Change-Id: I2e32dfa946d678fa59a290ddb53fc09387d59baf
Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>