Update patch set 4
Patch Set 4:

Ok I see now that it's defined (well more like "undefined") here https://github.com/lfit/releng-global-jjb/blob/master/jjb/lf-macros.yaml#L349

This is an interesting hack at JJB which doesn't scale. The way "lf-infra-sonar-with-prescan" is configured is that it expects a builder with a predefined name "lf-sonar-prescan" to be configured. However in JJB there can only ever be a single macro with the same name defined in a given JJB repo. So if 2 different projects require different prescan configuration it won't be possible with that macro. Perhaps this was overlooked, or perhaps the design was done this way with that in mind and the intent is to have a single prescan builder config for an entire JJB repo.

I'm not going to vote on this one and I'll leave it to LF folks whom are supporting this to approve.

Patch-set: 4
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>