Update patch set 4
Patch Set 4:

Bump, I'm sure you all are busy.  The ORAN go project teams want their stats to appear in SonarCloud, and this new template will make that possible.   But if you all feel this is a poor choice for global-jjb then I will lower it to ORAN jjb and use it there. Please advise.

Patch-set: 4