Update patch set 1
Patch Set 1:

This version pin of idna simply postpones the point in time when the failure will strike again. Does version 2.9 refuse to run in Python 3.5 or something that's being called by LF infrastructure scripts? I'm trying to ask, instead of a band-aid with absolutely zero comment explaining why the version was pinned, what's the right way to fix this?

Patch-set: 1
CC: Gerrit User 114 <114@f2a2391d-0403-4fa5-b183-4056999a8a09>