Update patch set 1
Patch Set 1:

DW it's great that you responded to Philippe so quickly, to get Acumos working again, and I don't to slow you down.  But IMHO this is the sort of change that adds technical debt & makes the LF infrastructure MORE fragile. Obviously I was too late, this is already merged.  But please consider.

Patch-set: 1