Update patch set 1
Patch Set 1:

DW

This change pinned python package idna.  But I see release jobs working fine WITHOUT the pin, for example in Acumos this job is using a slightly older version of global-jjb where release-job.sh does not yet have the pin of idna:

https://jenkins.acumos.org/job/portal-marketplace-release-merge/4/console

Why did the failure symptom go away?  Maybe the Python community reacted to the change in idna, and global-jjb does not have to add a pin?

If so, can this be reverted?

Patch-set: 1