Update patch set 1
Patch Set 1:

It's not needed but the macro expects it. I suspect you could either set it to a blank or "dummy" settings.xml that contains nothing.

Or create a new macro for sonarcloud that does not require settings.xml at all.

(I think in many cases though we were using settings.xml to pass credentials around regardless of maven or not maven projects)

Patch-set: 1
CC: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>