Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Please add a short release note about this version pin as a new file in releasenotes/notes/

ok, but I don't understand the file naming conventions there? Any documentation for how to
do this right?

I'm guessing this patch is a last ditch effort to unblock ODL CSIT jobs, but I wanted to get it
pushed to gerrit so I could share the link.

Patch-set: 1