Update patch set 2
Patch Set 2:

Right I'm not a member of OpenDaylight so my name doesn't appear in the share-with list. But just sharing it with me isn't enough, it should be 100% public if you're going to publish a link in the permanent git commit message.

Patch-set: 2
Reviewer: Gerrit User 114 <114@f2a2391d-0403-4fa5-b183-4056999a8a09>