Update patch set 3
Patch Set 3:

> Patch Set 2: Code-Review+1
> 
> (1 comment)
> 
> Aric thanks for the quick fix to this Acumos build problem, and for writing release notes.  I'm voting +1 because it clearly works.  But I'm only voting +1 because I deeply believe you should be pinning ALL the versions of the python components installed here. I believe this will prevent the next breakage when the maintainers of PyYAML, jsonschema, yamllint or what have you suddenly decide that no, Python 3.5 (or 3.6 or whatever) simply isn't good enough any more and break our world.  If the versions are pinned we are insulated against that kind of action, do I have that right?

Yeah, I mean, then I have to pin everyone at the lowest common denominator (python 3.5) and maybe that breaks things going forward. I find that if you pinn things they never get unpinned. I would rather fix things so that they work as expected in python 3.6, with as few exceptions for 2.5 as possible the the aim to drop support for python 3.5 (as the rest of the python community is doing) asap.

Patch-set: 3