Update patch set 2
Patch Set 2:

Yes this change preserves the ability of an upstream project/job definition to pass whatever value for cron on the Docker merge job is appropriate, including the empty string to disable that entirely.  This only changes the default value.

The docker world doesn't distinguish between merge and stage.  It only has merge.  So the merge job is triggered by Gerrit merge events, and *also* (I argue) should be triggered routinely.

If I'm only confusing the issue further with these comments I'll be glad to talk on the phone with you Anil at your convenience.  I'm home 24/7 now :/

Patch-set: 2