Update patch set 18
Patch Set 18:

> Patch Set 18:
> 
> Thanks Andrew.  I'll wait for another vote or two, and I don't like to submit my own change.
> 
> I'd also like to ask your opinion on this: as written now the change changes no behavior, projects must turn it on explicitly.  Alternately it could be on by default. But that could cause some angst when jjb submissions fail due to build-node entries in old/outdated files that nobody cares about.  What's your pref?

I gave it a +2 because it's a no-op for projects by default. I would be a lot more hesitant if it was on by default. I would rather that the projects self-elect to use it. After we get projects all onboard, we can flip to on by default, but at least this way we don't have to worry about breaking things that are currently "working" and allow us to properly test it before turning it on for a given project.

Patch-set: 18
Reviewer: Gerrit User 1 <1@f2a2391d-0403-4fa5-b183-4056999a8a09>