Add doc for jjb-verify build-node label check
Document the jjb-verify feature that checks build-node labels named
in JJB YAML files against nodes in cloud config files.
Change the feature's JJB configuration variables to share a prefix.
Improve the shell script to be robust to errors such as a suffix of
".cfg" or an external label of "" (just two double quotes).

Change-Id: I8beb5669531fe2b7f59a9bcf3706b27ede3965b9
Issue-ID: RELENG-2828
Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>