Update patch set 6
Patch Set 6: Code-Review-1

(7 comments)

I know it's a pain, but would you please consider biting the bullet and correcting the plural forms to be singular in the merge and verify-build templates? The packer jobs are not invoked THAT often.  I promise to help with the fix in Acumos and ORAN if you go that route.

Patch-set: 6
Reviewer: Gerrit User 114 <114@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-1