Update patch set 4
Patch Set 4:

(2 comments)

> Patch Set 2:
> 
> Hi Anil, I looked at job 339, I didn't see an error there. I think this job uses the script without the --user install invocation?
> 
> I was just questioning a minor point, trying to understand your comment in the release notes which seem to state that an invocation of "pip install --user .." errors out on lack of privs to write to /usr/local.  Sure, if you omit the --user argument then I expect pip to install to /usr/local, but that's a different situation entirely.

Hello Chris, The job 339 fixed the error, but was not the expected fix since we needed thoses lines, and some gaps as to why it was put in place, because of missing comments/info in the git history. Now there is some clarity and I have added a comment. Also the release notes and commit message is updated.

Patch-set: 4