Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> for my education as a python developer, please tell me, why would a pip --user install go anywhere  near /usr/local?

I think that is a red herring, and the broke due to some regression (probably feature) or other. I was unable to reproduce this.
In any case making a new venv fixes this, which is a common python trope.

Patch-set: 4
Reviewer: Gerrit User 12 <12@f2a2391d-0403-4fa5-b183-4056999a8a09>