Update patch set 7
Patch Set 7:

Thanks for fixing the docs.  But I think the docs should be a bit more clear.  The template never ever receives a list of platforms or templates.  This is the magic of JJB and its lists.  For example, you or I configure a list such as the example you cited in Acumos for this job:

    platforms:
      - centos-7
      - ubuntu-18.04

On this input, the jenkins-job script finds the first entry in the list and processes the template (creates a job) with platform set to value "centos-7".  Then it finds the second entry and reprocesses the template to create a job with platform set to value "ubuntu-18.04".

I'll give you another acumos example:

    stream:
      - master:
          branch: master
      - clio:
          branch: clio

Here we could argue whether the first keyword should be "streams".  But I think the singular form is right bcos that key (that configuration parameter) takes on exactly one value at one time, first "master" and second "clio".

Sorry for the long diatribe here. Hope this explains my dislike for the plural form.

Patch-set: 7
Reviewer: Gerrit User 114 <114@f2a2391d-0403-4fa5-b183-4056999a8a09>