Add .editorconfig file
EditorConfig is a universal editor configuration file supported by many
IDEs and Editors (although some might require installing a plugin) to
share consistent editor configuration with developers using different
editors.

This also updates the README.md file to remove the single item bullet
list as that was causing prettier to reformat the bullet item to insert
extra spacing between the '-' and first word in the list.

Ref: https://editorconfig.org/
Change-Id: I22605008837c567d9f6a62abfc85a195acc708c6
Signed-off-by: Thanh Ha <zxiiro@gmail.com>