Update patch set 4
Patch Set 4:

(1 comment)

> Patch Set 3:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2: Code-Review-1
> > > 
> > > (1 comment)
> > 
> > This throws an error in defaults.yaml is not read under .jjb-test/
> 
> Not sure what you mean here but I updated the patch to use lf_jjb_common and it appears to work fine for me locally. Maybe you forgot to update BOTH the Gerrit and GitHub versions of the job?

To be clear the reason we don't want to put those variables into defaults.yaml is because then it would be a change that would require ALL projects to suddenly have to add jjb-version and jjb-cache into their own defaults.yaml. That is a change in behaviour as we never required that before.

Patch-set: 4
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>