Update patch set 1
Patch Set 1:

Ok I finally tracked this issue down to https://review.opendev.org/#/c/731169/ in upstream JJB. Projects will hit this issue if they upgrade to JJB 3.5.0 which was released last week. The cause is that the comment-event-parameter-mode was not configurable in previous versions of JJB so defaulted to whatever value the plugin itself was using which appears to be PLAIN. This default is contrary to what's documented in the Jenkins ? help message that instead says BASE64 is the default.

The author of the JJB patch is defaulting to what the help message states rather than what the plugin defaults to. I think this is fine.

This patch explicitly "retains" the expected behaviour and thus does not require downstream projects to make any changes (other than upgrading to the JJB release that this patch ends up going to).

Patch-set: 1