Update patch set 1
Patch Set 1:

This might be a AWK version thing. Whether you are using GNU AWK or a different AWK. I think Macs have a different AWK than Linux systems for example.

Any fix shouldn't break on the version of AWK that's available in the builders we deploy so would be good to confirm this passes there.

Patch-set: 1
CC: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>