Migrate write-good hook to pre-commit
This is a final patch to remove our dependency on the Coala platform
and use only pre-commit moving forward.

This patch also resolves write-good lint issues that were not caught
by the Coala version of write-good. Not sure why that was but at least
with pre-commit we now have a chance to improve the docs here.

Issue: RELENG-2642
Signed-off-by: Thanh Ha <zxiiro@gmail.com>
Change-Id: Ic9a48b2964ff5c689a01f3fd611e69ef9525e055