Update patch set 4
Patch Set 4: Code-Review-2

Ok... my -2 is mostly because I don't want to see this merged until the database schema gets an update to handle this as well as the import scripts which will have to handle both with and without having this job status column added. Otherwise this generated data is just going to get dropped on the floor and when the schema and import script get updated it won't fix old data even if we have it for the import without a complete wipe of the database and reimport of all logs.

If you can get some changes up against https://gerrit.linuxfoundation.org/infra/admin/repos/tooling/job-costing that would be appreciated

Patch-set: 4
Reviewer: Gerrit User 1 <1@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-2