Run packer against a specific builder
As preparation to enable the ability to configure multiple builders
in a single packer template, we need to enable the ability in the
job to select which builder to run.

This patch sets up a new variable "packer-builder" which is used to
select the builder definition in the templates file.

Defaulting to "openstack" as that's the most common use case at the
moment.

Issue: RELENG-1052
Signed-off-by: Thanh Ha <zxiiro@gmail.com>
Change-Id: Ib33a00f38a0005af69d52e7697322a2e84f6c674