Fix parallel tox mode when using bool jjb cfg
JJB takes in the word "true" unquoted as a boolean and converts it to
the Python "True" boolean causing the bash script to ultimately receive
a string that starts with an uppercase letter. Since we are comparing
against a lowercase "true" we should ensure the string is lowercased
before comparison.

Signed-off-by: Thanh Ha <zxiiro@gmail.com>
Change-Id: If85c4411c73c18c931a339592a0dd352f574131a