Update patch set 1
Patch Set 1: Code-Review-1

What's to say that this detection path isn't going to go away in the future as well?

According to the cloud-init docs there should be a /run/cloud-init/instance-data.json file which has the "public" data available from the metadata service saved off.

My exploring of systems shows me that at least with CentOS 7 on OpenStack that doesn't exist (it does on an EC2 instance) that being said in all cases of my looking around there is a /run/cloud-init/result.json that has data in it like the following:

/run/cloud-init/result.json from an OpenStack instance:

> {
>   "v1": {
>    "datasource": "DataSourceOpenStack [net,ver=2]",
>    "errors": []
>  }
> }

/run/cloud-init/result.json from an EC2 instance:

> {
>  "v1": {
>   "datasource": "DataSourceEc2Local",
>   "errors": []
>  }
> }

So instead of relying on a possible knob that may change under us (again) let's use something that seems to be more stable.

Patch-set: 1
Reviewer: Gerrit User 1 <1@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-1