Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Looks like I'm too late for this one but why introduce yet another java-version variable just for Sonar?
> 
> The existing java-version variable should be sufficient to cover both uses. I cannot see a use case where one would want mvn and sonar to both use entirely different versions for Java.

It was introduced because that's actually the needed case here. ONAP has several projects that have dependencies that are not compatible with Java11 at this time but SonarCloud has deprecated Java8 and will remove the ability to do scans using it in a few months.

Patch-set: 2
Reviewer: Gerrit User 1 <1@f2a2391d-0403-4fa5-b183-4056999a8a09>