Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > Looks like I'm too late for this one but why introduce yet another java-version variable just for Sonar?
> > > 
> > > The existing java-version variable should be sufficient to cover both uses. I cannot see a use case where one would want mvn and sonar to both use entirely different versions for Java.
> > 
> > It was introduced because that's actually the needed case here. ONAP has several projects that have dependencies that are not compatible with Java11 at this time but SonarCloud has deprecated Java8 and will remove the ability to do scans using it in a few months.
> 
> huh, odd. Okay I guess that's fine although a bit inconvenient :)

I thought of just replacing the java-version with JDK11 before the scan begins but I also wanted to bring the flexibility of setting the 2nd JDK independently to the user's choice in case we need to run the scanner with an even newer JDK than 11 in future. FYI the projects not compatible with JDK11 are Portal and SDC (DCAE is in the process of migrating but they will not be as fast).

Patch-set: 2