Update patch set 3
Patch Set 3:

I have issues with this. It's codifying the use of master as the primary / development branch. Given all the industry talk about moving off of using 'master' to 'main' or something else I think that if 'latest' is desired it _needs_ to be set to latest and any stream config _needs_ to set it to that stream.

Patch-set: 3
Reviewer: Gerrit User 1 <1@f2a2391d-0403-4fa5-b183-4056999a8a09>