Preliminary jcasc yaml generator
generator that takes values from a projects builder/jenkins-config/*.cfg
files and creates a yaml file compatible with jcasc

Idea is to replace the groovy scripts in this same directory
with something less brittle.

A few problems, we can't validate the schema.
as none has been written.

so loading whatever this generates will be risky.

example output:
https://gist.github.com/Aricg/4366e4e54e6ab63905ecb7aaac56646f

Issue-Id: RELENG-3035
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: I8fbb6a49f8d13ba36c32b30469410c3a69c0f5f0