Update patch set 1
Patch Set 1:

This still needs docs and changelog, but first, there's an uncertain implementation detail. I thought I found a way to scan a Github organization in the RTD jobs, but those jobs seem specifically made for Gerrit (though there's a GH version, the shell script that is being called expects a Gerrit env var to be present, and does no error handling, so it doesn't look like GH compatibility was ever truly implemented).

I've been doing some experimenting with org-level webhooks in Github, but haven't had any success so far. The Github trigger doesn't seem to respond to the hooks. We may need to define this job (possibly as part of the ci-man jobs) for each repo that is expected to have a Jenkinsfile. But if there are any more alternatives to try out, I'm open to looking into them.

Patch-set: 1
Reviewer: Gerrit User 1 <1@f2a2391d-0403-4fa5-b183-4056999a8a09>