Add new parameter to the build-discarder
Adding number of builds to keep parameter to the build-discarder macro
for FD.io. The FD.io retention for builds is based off the number of
builds as opposed to the number of days to keep due to the number of
builds that are run on a given day for a single job.

Signed-off-by: Vanessa Rene Valderrama <vvalderrama@linuxfoundation.org>
Change-Id: I347a26b41cd98a54293430f3ee367a962f49418e