Remove the -only CLI option for packer build
The CR 90dd675df482a8 that was submitted to work with
common-packer 0.7.0 does not work as expected and fails
all the packer jobs due to a bug in the packer build CLI.

There is a bug when using `-only` option packer build CLI
does not set the cloud provider name/type as per the documentation.
The alternative is to use an environment variable that is passed
through the packer job and can be reference in the packer
templates.

Change-Id: I97bc479ddac9a8f1f0dc83201735b0a084a8b3e4
Ref: https://gerrit.linuxfoundation.org/infra/c/releng/common-packer/+/66528
Issue: GH https://github.com/hashicorp/packer/issues/10495
Issue: JIRA IT-21300
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>